+++ src/utils/engine.py
@@ -29,7 +29,7 @@ def train_fn(data_loader, model, optimizer, device):
         optimizer.zero_grad()
         encoded, decoded = model(embeds)
 
-        loss = loss_fn(embeds.type(torch.float), decoded.type(torch.float))
+        loss = loss_fn(embeds, decoded)
         loss.backward()
         loss_tot += loss.item()
         # print(loss.item())
@@ -61,7 +61,7 @@ def eval_fn(data_loader, model, device):
 
             encoded, decoded = model(embeds)
 
-            loss = loss_fn(embeds.type(torch.float), decoded.type(torch.float))
+            loss = loss_fn(embeds, decoded)
             loss_tot += loss.item()
     epoch_loss = loss_tot/len(data_loader)
